requirements.txt
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
feedparser==6.0.11
jinja2==3.1.4
python-telegram-bot==20.7
pytz==2024.1
